messages d'erreurs retouches
