+++ only_colors.cub
@@ -7,7 +7,8 @@ EA      ./textures/east.xpm
 C 125,12,14
 F 14,14,14
 
-11111111111
-100N0011111111111111111111111111111
-10000W010000111111
-11111111111111 11
+111111111111111111111111111111
+111111110000011111111
+111111100000010000111
+      11000000N0000010
+1111111111111111111111111111111
