+++ srcs/parsing/map_lines_parser/core_map_parser.c
@@ -16,19 +16,17 @@ t_map_status	core_map_parser(t_map_data *map)
 {
 	if (check_if_open_map(map) == INVALID_MAP)
 	{
-		ft_dprintf(STDERR_FILENO, "Error\nInvalid map.\n");
-		return (INVALID_MAP); //exit? free? msg erreur perso?
+		return (INVALID_MAP);
 	}
 	if (check_if_invalid_core_map(map) == INVALID_MAP)
 	{
-		printf("Salut\n");
-		ft_dprintf(STDERR_FILENO, "Error\nInvalid map.\n");
-		return (INVALID_MAP); //exit? free? msg erreur perso?
+		ft_dprintf(STDERR_FILENO, "Error\nInvalid map, please try again.\n");
+		return (INVALID_MAP);
 	}
 	if (is_valid_number_of_players(map->players_number) == false)
 	{
 		ft_dprintf(STDERR_FILENO, "Error\nInvalid number of players.\n");
-		return (INVALID_MAP); //exit? free?
+		return (INVALID_MAP);
 	}
 	return (VALID_MAP);
 }
